CCallHelp.jl

Documentation for CCallHelp.jl

CCallHelp.free_ptr — Function

free_ptr(ptr; free=true)

Free the given ptr if not C_NULL.

Arguments

ptr the C pointer to be freed

free a conditional flag to prevent an if at the call site

CCallHelp.string_from_ptr — Function

string_from_ptr(ptr::Ptr{Cuchar}; free=true)

Create a Julia String from a memory location a byte at a time.

Arguments

ptr pointer to use

free flag to determine if the memory should be freed once copied

CCallHelp.vec_string_from_ptr — Function

vec_string_from_ptr(ptr_ptr_uchar; free=true)

Create a Vector{String} from the supplied list of uchar*

Arguments

ptr the pointer to the list

free flag to determine if the memory of both the strings and the list should be freed once copied